Backend/requirements.txt
fastapi
uvicorn
asyncpg
websockets
python-dotenv
psutil 
redis
python-multipart